pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "paradox-solutions-llm"
version = "0.3.2"
description = "Paradox Solutions LLM — local AI research assistant (Windows-friendly)"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.11"
classifiers = [
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "httpx>=0.27.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "chromadb>=0.5.0",
  "psutil>=6.0.0",
  "trafilatura>=1.6.0",
  "pypdf>=5.0.0",
  "python-multipart>=0.0.9",
  "pyyaml>=6.0.0",
  "ddgs>=9.0.0",
  "beautifulsoup4>=4.12.0",
  "cryptography>=42.0.0",
]

[project.optional-dependencies]
discord = ["discord.py>=2.3.0"]
eeg = [
  "mne>=1.6.0",
  "pyedflib>=0.1.37",
  "numpy>=1.26.0",
  "scipy>=1.12.0",
  "matplotlib>=3.8.0",
  "pandas>=2.1.0",
  "mne-icalabel>=0.7.0",
  "autoreject>=0.4.3",
  "mne-connectivity>=0.6.0",
  "plotly>=5.18.0",
]
all = [
  "paradox-solutions-llm[discord,eeg]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["hexnode*"]
